+++ src/conf/rpc.rs
@@ -3,14 +3,12 @@ use smart_default::*;
 
 use std::time::Duration;
 
-use crate::conf::duration;
-
 /// Server represents [`Server`] configuration section.
-#[derive(Clone, Debug, Deserialize, Serialize, SmartDefault)]
+#[derive(Clone, Debug, Serialize, Deserialize, SmartDefault)]
 pub struct Rpc {
     /// Timeout for [`WsSession`] to wait ping message from [`Web Client`].
-    #[serde(serialize_with = "duration::serialize")]
-    #[serde(deserialize_with = "duration::deserialize")]
+    #[serde(deserialize_with = "serde_humantime::deserialize")]
+    #[serde(serialize_with = "crate::utils::duration::serialize")]
     #[default(Duration::from_secs(10))]
     pub idle_timeout: Duration,
 }
